MULE-18102: Solution improvements
